requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.10.1
python-multipart>=0.0.9

# Document generation
python-docx>=1.1.2
lxml>=5.3.0
Pillow>=10.0.0

# MCP Server
mcp>=1.2.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.2
